requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
djoser djangorestframework-simplejwt==4.7.2
Pillow==9.0.0
django-filter==23.2
